MAGIC mode is running on microcontroller

+++ microcontroller/cw521/magic_bachelorarbeit/src/salsa20.c
@@ -107,7 +107,7 @@ void salsa20_expansion(uint32_t state_in[16], uint32_t state_out[16], uint32_t k
 }
 
 
-void salsa20_encryption_decryption(uint32_t key[8], uint32_t nonce[2], uint32_t *message_in, uint32_t *message_out, uint64_t message_length) {
+void salsa20_encryption_decryption(uint32_t key[8], uint32_t nonce[2], uint32_t *message_in, uint32_t *message_out, int message_length) {
     uint32_t state_in[16], state_out[16];
 
     uint32_t nonce_with_position[4];
